Kabinets-js first stable release tested as host library
