+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kabinets-react",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Kabinets-react is the simplest Global State Management for React.js",
   "main": "src/module.js",
   "scripts": {
@@ -25,6 +25,9 @@
     "url": "https://github.com/andreidim/cabinets-react.js/issues"
   },
   "homepage": "https://github.com/andreidim/cabinets-react.js#readme",
+  "peerDependencies": {
+   "react": ">=16.8.0"
+},
   "dependencies": {
     "kabinets": "^0.0.1"
   },
@@ -37,8 +40,8 @@
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "jest": "^26.6.3",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
     "react-test-renderer": "^16.14.0",
     "webpack": "^5.21.1",
     "webpack-cli": "^4.5.0"
